Extract dokuwiki\Search\Indexer class
Not sure why Doku_Indexer caused phpcs to complain about the class name
not being in PascalCase, but Doku_Handler didn't. 🤷

The namespace and new class name was selected to be compatible with the
upcoming changes in #2943. This should hpopefully reduce the overall
hassle of touching the same code base.
