+++ _test/phpcs.xml
@@ -149,6 +149,7 @@
         <exclude-pattern>*/inc/farm.php</exclude-pattern>
         <exclude-pattern>*/inc/fulltext.php</exclude-pattern>
         <exclude-pattern>*/inc/pluginutils.php</exclude-pattern>
+        <exclude-pattern>*/inc/indexer.php</exclude-pattern>
         <exclude-pattern>*/inc/Mailer.class.php</exclude-pattern>
         <exclude-pattern>*/doku.php</exclude-pattern>
         <exclude-pattern>*/install.php</exclude-pattern>
