+++ _test/tests/inc/indexer_rename.test.php
@@ -3,7 +3,7 @@
  * Test cases for the Doku_Indexer::renamePage and Doku_Indexer::renameMetaValue methods
  */
 class indexer_rename_test extends DokuWikiTest {
-    /** @var Doku_Indexer $indexer */
+    /** @var \dokuwiki\Search\Indexer $indexer */
     private $indexer;
 
     private $old_id = 'old_testid';
