+++ inc/deprecated.php
@@ -563,3 +563,8 @@ class Subscription {
         $manager->notifyAddresses($data);
     }
 }
+
+/**
+ * @deprecated 2019-12-29 use \dokuwiki\Search\Indexer
+ */
+class Doku_Indexer extends \dokuwiki\Search\Indexer {};
